feat: Add DataFixtures for App and Course entities
This commit adds DataFixtures for the App and Course entities. The AppFixtures class is responsible for loading initial data into the database, while the CourseFixtures class generates dummy data for the Course entity and its related entities (Trainer and CourseModule). These fixtures will be used for testing and populating the database with sample data.