feat: Add Resource, CourseCohort, TraineeResource, and FeedbackCategory entities
This commit adds the Resource, CourseCohort, TraineeResource, and FeedbackCategory entities to the application. These entities represent different aspects of the application's data model and will be used for storing and managing resources, course cohorts, trainee resources, and feedback categories. The entities are defined with the necessary properties and relationships to support the application's functionality.