feat: Add migrations for trainee documents and notification fields
This commit adds migrations for trainee documents and notification fields. It includes changes to the trainee table to modify the "documents" column from VARCHAR(255) to LONGTEXT and changes to the notification table to add the "text" column as VARCHAR(255) and modify the "title" column to "from" as VARCHAR(150). These migrations are necessary to update the database schema and accommodate the new requirements for storing trainee documents and notification details.