feat: Update download method to use TrainerRepository
This commit updates the download method in the HomeController to include the TrainerRepository as a parameter. This change allows the method to retrieve the path for the downloaded file from the "homeworks/trainers" directory instead of using the user's identifier. This modification improves the accuracy of file retrieval and ensures that the correct file is downloaded for each trainer.