feat: Add getAddress method to Prospect entity
This commit adds a new method called getAddress to the Prospect entity. The getAddress method returns a string representation of the prospect's address by concatenating the number, street, and additional address fields. This addition enhances the functionality of the Prospect entity by providing a convenient way to retrieve the full address of a prospect.