feat: Update module and course templates with translated text
This commit updates the module.html.twig and course.html.twig templates to use translated text for headings and content. By using translation keys and the trans filter, we ensure that the templates display the correct text based on the user's language preference. This modification enhances the internationalization of the application and improves the user experience.